Many changes are made to make the solution working
