+++ index.html
@@ -3,29 +3,37 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Serverless Application with an API to ChatGPT Interface</title>
-    <link rel="stylesheet" href="style.css" />
-    <script src="script.js" defer></script>
+    <title>
+      CIG -Serverless Application Integration with ChatGPT - Example2
+    </title>
+    <link rel="stylesheet" href="$BASE_URL$/style.css" />
+    <script src="$BASE_URL$/script.js" defer></script>
   </head>
+
   <body>
     <div class="banner">
-      <img src="page_top_banner.jpg" alt="Top Banner" />
+      <img src="$BASE_URL$/page_top_banner.jpg" alt="Top Banner" />
     </div>
 
-    <h1>Serverless Application with an API to ChatGPT Interface</h1>
+    <h1>CIG - Serverless Application with an API to ChatGPT Interface</h1>
 
     <form id="locationForm">
       <label for="location">Choose a location:</label>
       <select id="location" name="location">
-        <option value="delhi">Delhi</option>
-        <option value="mumbai">Mumbai</option>
-        <option value="bengaluru">Bengaluru</option>
-        <option value="chennai">Chennai</option>
-        <option value="hyderabad">Hyderabad</option>
-        <option value="kolkata">Kolkata</option>
+        <option value="Bengaluru">Bengaluru</option>
+        <option value="Chennai">Chennai</option>
+        <option value="Chandigarh">Chandigarh</option>
+        <option value="Delhi">Delhi</option>
+        <option value="Hyderabad">Hyderabad</option>
+        <option value="Kolkata">Kolkata</option>
+        <option value="Lucknow">Lucknow</option>
+        <option value="Mumbai">Mumbai</option>
+        <option value="Pune">Pune</option>
+        <option value="Thiruvananthapuram">Thiruvananthapuram</option>
       </select>
       <button type="submit">Submit</button>
     </form>
+    <hr />
 
     <h2>The details of location are:</h2>
     <div id="locationDetails"></div>
