+++ lambda_Landing_Page.py
@@ -1,6 +1,11 @@
 #------------------------------------------------------------------------------------
-# Created by Rama Subramanyam based on video at youtube at https://www.youtube.com/watch?v=Lf98s3NczBE
+# Created by Rama Subramanyam based on video at youtube at https://www.youtube.com/watch?v=XX6ADwJBGgY&t=95s
 # This is the second example in the series to learn implementation serverless app at AWS
+# Following are the features of this lambda function
+# a) This is lambda_function when access by suer using function url,
+#    which reads index.html from S3 return the same to user screen.  
+# b) The index.html page has a form for the user to select the location and submit
+#    When submitted, it calls another lambda function.  
 # 
 # date : May 2025
 # usage: over ride default variable values
@@ -11,7 +16,7 @@
 # aws_acount_id = "       "
 # tags = {
 #   "terraform" = "true"
-#   "project" = "Lambda-OpenAI-Interface-Example1"
+#   "project" = "Lambda-OpenAI-Interface-Example2"
 #   "contact" = "ecirams@gmail.com"
 # }
 # aws_region = "us_east_1"
@@ -24,42 +29,26 @@
 # debug_mode = true
 # lambda_siz = 256
 # lambda_timeout = 400
+###################################################
 
-####################################################
-# Lambda Environment Variable Settings
-# API_KEY = "<openai-key>" 
-####################################################
+import boto3
 
+s3_client = boto3.client('s3')
 
-import openai
-import os
+def lambda_handler(event,context):
+    s3_Bucket_Name = "TO-REPLACE-ME"
+    s3_File_Name = "index.html"
+    base_url = "TO-REPLACE-ME"
 
-
-def lambda_handler(event, context):
-    #Get Environment variable  
-    openai.api_key = os.getenv("API_KEY")
-
-    #Set value for location for which you would like to display interesting fact to user
-    location = "New York, NY"
-    
-    messages = [
-        {"role": "system", "content": "You are a helpful assistant."},
-        {"role": "user", "content": f"Tell me interesting fact about {location} in less than 200 characters "}
-    ]
-
-    # chat = openai.Completion.create(model="gpt-3.5-turbo", messages = messages)
-    chat = openai.chat.completions.create(model="gpt-3.5-turbo", messages = messages)
-
-    gpt_response = chat.choices[0].message.content
-
-    html_output = f"<html><body><h1>Interesting Fact about the place {location}</h1><p>{gpt_response}</p></body></html>"
-
-    print("gpt_response:",gpt_response)
+    object = s3_client.get_object(Bucket=s3_Bucket_Name, Key=s3_File_Name)
+    body = object['Body']
+    html_output = body.read().decode('utf-8')
+    html_output = html_output.replace ('$BASE_URL$',base_url)
 
     return {
-        'statusCode': 200,
+        'statusCode' : 200,
         'body': html_output,
-        'headers': {
-            'Content-Type': 'text/html'
+        'headers' : {
+            'Content-Type' : 'text/html',
         }
     }
